+++ NEWS.md
@@ -12,16 +12,20 @@
 - [ ] Verify Engine compatibility with all major engines
 - [ ] Handle diff-in-diff and fixed effects formula notation
 - [ ] Add graphics: gray scale of the original model, with the minimal set visibly differentiated.
-- [ ] Add options for LaTeX and Kable output
+- [ ] Add options for LaTeX output
 - [ ] Add export_report functionality, which will print a publication-grade 2-page robustness check with a single, simple call (for robustness check use)
 - [ ] Add a simple "covariate sensitivity" line to report, indicating if (i) the original model is statistically significant; (ii) the minimal or canonical set is statistically insignificant
 - [ ] Run `dag_assist` on a bunch of complex DAGS to refine the sorting algorithm in `compare.R`.
 - [ ] Change `dag_assist` to `DAGassist`.
 - [ ] Add `imply`, default TRUE, logical parameter to disable adding variables to formula/table based on DAG relationships.
+- [ ] Get rid of the engine args param; it is obsolete
 
 ## UNRELEASED
 
-### 0.2.0.9100 - 2025-08-26
+### 0.2.0.9100 - 2025-08-28
+
+#### Added
+- report_latex.R provides LaTeX output functionality
 
 #### Changed
 - **Added some pretty colors to console report output**